API/requirements.txt
django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.4.0